pyproject.toml
[project]
name = "loan-agent"
version = "0.1.0"
description = "A 15-step loan application journey using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain>=0.3",
    "langchain-google-genai>=4.0",
    "langsmith>=0.1",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "streamlit>=1.30",
    "pytest>=8.0",
]
